+++ android/src/main/cpp/graphics/shader/ColorLineGroup2dShaderOpenGl.h
@@ -53,6 +53,6 @@ protected:
     //const int sizeGapColorValues = 4;
     //const int maxNumDashValues = 4;
    // const int sizeDashValues = maxNumDashValues + 1;
-    const int sizeLineValues = 19;//sizeStyleValues + sizeColorValues + sizeGapColorValues + sizeDashValues + 1;
+    const int sizeLineValues = 21;//sizeStyleValues + sizeColorValues + sizeGapColorValues + sizeDashValues + 1;
     const int sizeLineValuesArray = sizeLineValues * maxNumStyles;
 };