+++ bridging/android/jni/map/layers/line/NativeLineStyle.cpp
@@ -24,13 +24,15 @@ auto NativeLineStyle::fromCpp(JNIEnv* jniEnv, const CppType& c) -> ::djinni::Loc
                                                            ::djinni::get(::djinni::F32::fromCpp(jniEnv, c.width)),
                                                            ::djinni::get(::djinni::List<::djinni::F32>::fromCpp(jniEnv, c.dashArray)),
                                                            ::djinni::get(::djinni_generated::NativeLineCapType::fromCpp(jniEnv, c.lineCap)),
-                                                           ::djinni::get(::djinni::F32::fromCpp(jniEnv, c.offset)))};
+                                                           ::djinni::get(::djinni::F32::fromCpp(jniEnv, c.offset)),
+                                                           ::djinni::get(::djinni::Bool::fromCpp(jniEnv, c.dotted)),
+                                                           ::djinni::get(::djinni::F32::fromCpp(jniEnv, c.dottedSkew)))};
     ::djinni::jniExceptionCheck(jniEnv);
     return r;
 }
 
 auto NativeLineStyle::toCpp(JNIEnv* jniEnv, JniType j) -> CppType {
-    ::djinni::JniLocalScope jscope(jniEnv, 10);
+    ::djinni::JniLocalScope jscope(jniEnv, 12);
     assert(j != nullptr);
     const auto& data = ::djinni::JniClass<NativeLineStyle>::get();
     return {::djinni_generated::NativeColorStateList::toCpp(jniEnv, jniEnv->GetObjectField(j, data.field_color)),
@@ -41,7 +43,9 @@ auto NativeLineStyle::toCpp(JNIEnv* jniEnv, JniType j) -> CppType {
             ::djinni::F32::toCpp(jniEnv, jniEnv->GetFloatField(j, data.field_width)),
             ::djinni::List<::djinni::F32>::toCpp(jniEnv, jniEnv->GetObjectField(j, data.field_dashArray)),
             ::djinni_generated::NativeLineCapType::toCpp(jniEnv, jniEnv->GetObjectField(j, data.field_lineCap)),
-            ::djinni::F32::toCpp(jniEnv, jniEnv->GetFloatField(j, data.field_offset))};
+            ::djinni::F32::toCpp(jniEnv, jniEnv->GetFloatField(j, data.field_offset)),
+            ::djinni::Bool::toCpp(jniEnv, jniEnv->GetBooleanField(j, data.field_dotted)),
+            ::djinni::F32::toCpp(jniEnv, jniEnv->GetFloatField(j, data.field_dottedSkew))};
 }
 
 } // namespace djinni_generated