+++ ios/graphics/Texture/TextureHolder.swift
@@ -28,7 +28,15 @@ public class TextureHolder: NSObject {
         super.init()
     }
 
-    convenience init(_ url: URL, textureUsableSize: TextureUsableSize? = nil) throws {
+    deinit {
+        let texture = self.texture
+        DispatchQueue.global(qos: .utility).async {
+            // access texture to make sure it gets deallocated asynchron
+            _ = texture.width
+        }
+    }
+
+    public convenience init(_ url: URL, textureUsableSize: TextureUsableSize? = nil) throws {
         let options: [MTKTextureLoader.Option: Any] = [
             MTKTextureLoader.Option.SRGB: NSNumber(booleanLiteral: false),
         ]