+++ shared/src/map/layers/tiled/vector/Tiled2dMapVectorLayerParserHelper.cpp
@@ -298,7 +298,9 @@ Tiled2dMapVectorLayerParserResult Tiled2dMapVectorLayerParserHelper::parseStyleJ
                     parser.parseValue(val["paint"]["line-blur"]),
                     parser.parseValue(val["layout"]["line-cap"]),
                     parser.parseValue(val["paint"]["line-offset"]),
-                    blendMode
+                    blendMode,
+                    parser.parseValue(val["paint"]["line-dotted"]),
+                    parser.parseValue(val["paint"]["line-dotted-skew"])
             );
             auto layerDesc = std::make_shared<LineVectorLayerDescription>(
                     val["id"],