+++ CHANGELOG.md
@@ -1,5 +1,35 @@
 # Changelog for Open Mobile Maps
 
+
+## Version 2.2.0
+- Adds support for `{bbox-epsg-3857}` placeholder in URL template
+
+## Version 2.1.0 (08.04.2024)
+- Adds dotted line style (line-dotted boolean and an optional skewFactor, defaulting to 1.0 and denoting the stretch factor)
+- Fixes style for short dashes
+- Fixes line touch area
+- Fix map description issues in concurrent settings
+- Added a Helper to create geojson strings from GeoJsonPoints
+
+## Version 2.0.7 (29.04.2024)
+- adds performance logger (disabled by default)
+- fixes a bug in the android striped pattern shader
+- fixes a bug in geojson hole simplification
+
+## Version 2.0.6 (10.04.2024)
+- log exception when a font can not be loaded
+
+## Version 2.0.5 (08.04.2024)
+- Expose thread_pool_callbacks in scheduler interface
+- call onRemoved on layerInterface before the mapScene is destroyed
+
+## Version 2.0.4 (22.03.2024)
+- Changes in symbol vector map collision detection: Run collision detection when zoom changes. This change ensures that icons never collide but also carries a performance burden.
+
+## Version 2.0.3 (19.03.2024)
+- Fixes thread safety in iOS shader
+- Fixes coordinate conversion from EPSG:4326 to EPSG:3857
+
 ## Version 2.0.2 (12.03.2024)
 - Fixes a bug in iOS Polygon Pattern Shader
 - Fixes a bug in camera restricted bounds calculations